As titled, there is a slightly improvements in the performances, any suggestion?

Before:
```
lookup static route x 21,752,581 ops/sec ±0.75% (90 runs sampled)
lookup dynamic route x 707,120 ops/sec ±1.14% (91 runs sampled)
lookup long static route x 2,894,137 ops/sec ±7.11% (86 runs sampled)
find static route x 31,506,022 ops/sec ±1.12% (87 runs sampled)
find dynamic route x 705,175 ops/sec ±0.85% (88 runs sampled)
find long static route x 3,994,827 ops/sec ±10.60% (83 runs sampled)
```

After:
```
lookup static route x 21,755,241 ops/sec ±0.83% (89 runs sampled)
lookup dynamic route x 717,386 ops/sec ±1.07% (90 runs sampled)
lookup long static route x 2,768,482 ops/sec ±6.72% (86 runs sampled)
find static route x 31,849,514 ops/sec ±0.79% (89 runs sampled)
find dynamic route x 737,994 ops/sec ±1.01% (87 runs sampled)
find long static route x 4,296,578 ops/sec ±10.29% (91 runs sampled)
```